requirements.txt
django
djangorestframework
dj-rest-auth
psycopg2-binary
pillow
python-decouple
